#6968 Project metadata folder create on demand (fix).
Former-commit-id: 092d76477890adc15b496526254e548841183092

+++ plugins/org.jkiss.dbeaver.registry/src/org/jkiss/dbeaver/registry/DataSourceRegistry.java
@@ -675,7 +675,9 @@ public class DataSourceRegistry implements DBPDataSourceRegistry {
                     IOUtils.makeFileBackup(plainConfigFile);
 
                     if (localDataSources.isEmpty()) {
-                        configFile.delete(true, false, monitor.getNestedMonitor());
+                        if (configFile.exists()) {
+                            configFile.delete(true, false, monitor.getNestedMonitor());
+                        }
                     } else {
                         DataSourceSerializer serializer;
                         if (project.getFormat() == ProjectMetadata.ProjectFormat.LEGACY) {
@@ -683,6 +685,7 @@ public class DataSourceRegistry implements DBPDataSourceRegistry {
                         } else {
                             serializer = new DataSourceSerializerModern(this);
                         }
+                        project.getMetadataFolder(true);
                         serializer.saveDataSources(
                             monitor,
                             origin,
